Feat: suggestions

+++ server/controllers/suggestion.py
@@ -4,7 +4,7 @@ from psycopg2.extras import RealDictCursor
 
 def create_suggestion(account_id, post, status_id, category_id, slug, suggestion_title):
     """ Create new account_id into  the acount table """
-
+    status_id = 1
     sql = """INSERT INTO suggestion (account_id, post, status_id, category_id, slug, suggestion_title)
              VALUES(%s, %s, %s, %s, %s, %s) RETURNING id;"""
     
@@ -15,13 +15,10 @@ def create_suggestion(account_id, post, status_id, category_id, slug, suggestion
             with  conn.cursor() as cur:
                 # execute the INSERT statement
                 cur.execute(sql, (account_id, post, status_id, category_id, slug, suggestion_title))
-
-                # get the generated id back                
+            
                 rows = cur.fetchone()
                 if rows:
                     response = rows
-
-                # commit the changes to the database
                 conn.commit()
     except (Exception, psycopg2.DatabaseError) as error:
         print(error)    
@@ -41,13 +38,10 @@ def create_suggestion_response(account_id, res, suggestion_id):
             with  conn.cursor() as cur:
                 # execute the INSERT statement
                 cur.execute(sql, (account_id, res, suggestion_id))
-
-                # get the generated id back                
+            
                 rows = cur.fetchone()
                 if rows:
                     response = rows
-
-                # commit the changes to the database
                 conn.commit()
     except (Exception, psycopg2.DatabaseError) as error:
         print(error)    
@@ -63,15 +57,12 @@ def fetch_suggestion(id):
     try:
         with  connection as conn:
             with  conn.cursor(cursor_factory=RealDictCursor) as cur:
-                # execute the UPDATE statement
-                cur.execute(query, (int(id), ))
 
-                # get the generated id back                
+                cur.execute(query, (int(id), ))
+            
                 rows = cur.fetchone()
                 if rows:
                     response = rows
-
-                # commit the changes to the database
                 conn.commit()
     except (Exception, psycopg2.DatabaseError) as error:
         response = error
@@ -80,7 +71,7 @@ def fetch_suggestion(id):
     
 # fetch all users
 def fetch_suggestions():
-    query = """SELECT * FROM suggestion JOIN account on account_id = account.id JOIN category on category_id = category.id JOIN status on status_id = status.id  ORDER BY suggestion_time;"""
+    query = """SELECT suggestion.*, suggestion.id AS suggestion_id, account.*, account.id AS account_id FROM suggestion JOIN account on account_id = account.id JOIN category on category_id = category.id JOIN status on status_id = status.id  ORDER BY suggestion_time;"""
     
     response = None
 
@@ -94,16 +85,37 @@ def fetch_suggestions():
                 rows = cur.fetchall()
                 if rows:
                     response = rows
-                # commit the changes to the database
                 conn.commit()
     except (Exception, psycopg2.DatabaseError) as error:
         response = error
     finally:
         return response
 
-# update user
-def edit_suggestion(id, post, category_id, status_id):
-    query = """UPDATE account SET (post=%s,category_id=%s,status_id=%s) WHERE id = %s RETURNING title
+
+# fetch suggestion responses
+def fetch_suggestion_response(id):
+    query = """SELECT suggestion_response.*, suggestion_response.id AS response_id, account.email, account.username, account.lastname, account.is_admin, account.is_staff, account.profile, account.user_title_id, user_title.user_title FROM suggestion_response JOIN account ON account_id = account.id JOIN user_title ON user_title_id = user_title.id WHERE suggestion_id=%s ;"""
+    
+    response = None
+
+    try:
+        with  connection as conn:
+            with  conn.cursor(cursor_factory=RealDictCursor) as cur:
+
+                cur.execute(query, (int(id), ))
+            
+                rows = cur.fetchall()
+                if rows:
+                    response = rows
+                conn.commit()
+    except (Exception, psycopg2.DatabaseError) as error:
+        response = error
+    finally:
+        return response
+
+def edit_suggestion(title, post, category_id, account_id, suggestion_id):
+
+    query = """UPDATE suggestion SET suggestion_title = %s,post = %s,category_id = %s WHERE id = %s AND account_id = %s RETURNING suggestion_title
     ;"""
     
     response = None
@@ -111,39 +123,81 @@ def edit_suggestion(id, post, category_id, status_id):
     try:
         with  connection as conn:
             with  conn.cursor() as cur:
-                # execute the UPDATE statement
-                cur.execute(query, (post, category_id, status_id, int(id)))
 
-                # get the generated id back                
+                cur.execute(query, (title, post, category_id, suggestion_id, account_id))            
                 rows = cur.fetchone()
                 if rows:
                     response = rows[0]
-
-                # commit the changes to the database
                 conn.commit()
     except (Exception, psycopg2.DatabaseError) as error:
         response = error
     finally:
         return response
     
-# update user
-def delete_suggestion(id):
-    query = """DELETE FROM suggestion WHERE id=%s RETURNING id;"""
+
+def delete_suggestion(suggestion_id, account_id):
+    print("ID ID: ", account_id)
+    print("SUGGESTION ID: ", suggestion_id)
+    query = """DELETE FROM suggestion WHERE id = %s AND account_id = %s RETURNING id;"""
     
     response = None
 
     try:
         with  connection as conn:
             with  conn.cursor() as cur:
-                # execute the UPDATE statement
-                cur.execute(query, (int(id), ))
 
-                # get the generated id back                
+                cur.execute(query, (suggestion_id, account_id,) )
+            
                 rows = cur.fetchone()
                 if rows:
                     response = rows[0]
+                conn.commit()
+    except (Exception, psycopg2.DatabaseError) as error:
+        response = error
+    finally:
+        return response
+    
+    
+def create_suggestion_comment(account_id, comment, suggestion_id):
+    """ Create new account_id into  the acount table """
+    print("ACCOUNT_ ID: ", account_id)
+    print("COMMENT: ", comment)
+    print("SUGGESTION ID: ", suggestion_id)
+    sql = """INSERT INTO suggestion_comment (account_id, comment, suggestion_id)
+             VALUES(%s, %s, %s) RETURNING id;"""
+    
+    response = None
 
-                # commit the changes to the database
+    try:
+        with  connection as conn:
+            with  conn.cursor() as cur:
+                # execute the INSERT statement
+                cur.execute(sql, (account_id, comment, suggestion_id))
+            
+                rows = cur.fetchone()
+                if rows:
+                    response = rows
+                conn.commit()
+    except (Exception, psycopg2.DatabaseError) as error:
+        print(error)    
+    finally:
+        return response
+
+# fetch suggestion responses
+def fetch_suggestion_comments(id):
+    query = """SELECT suggestion_comment.*, suggestion_comment.id AS comment_id, account.email, account.username, account.lastname, account.is_admin, account.is_staff, account.profile, account.user_title_id, user_title.user_title FROM suggestion_comment JOIN account ON account_id = account.id JOIN user_title ON user_title_id = user_title.id WHERE suggestion_id=%s ;"""
+    
+    response = None
+
+    try:
+        with  connection as conn:
+            with  conn.cursor(cursor_factory=RealDictCursor) as cur:
+
+                cur.execute(query, (int(id), ))
+            
+                rows = cur.fetchall()
+                if rows:
+                    response = rows
                 conn.commit()
     except (Exception, psycopg2.DatabaseError) as error:
         response = error
